1_Demo_Agents/requirements.txt
# DSPy vs LangGraph Benchmark - Requirements
# ==========================================

# Core frameworks
dspy>=2.5.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Yahoo Finance tools
yfinance>=0.2.40

# Utilities
python-dotenv>=1.0.0

# For benchmarking (optional)
pandas>=2.0.0
matplotlib>=3.7.0